feature #15 user -> post로 도메인 맞게 리팩토링
