+++ src/main/java/com/search/instagramsearching/controller/UsersController.java
@@ -23,12 +23,6 @@ public class UsersController {
         return ResponseDto.success(usersService.searchUsers(keyword, pageable));
     }
 
-    // 유저 profileName으로 유저의 게시글 조회하기 API
-    @GetMapping("/user/{userSid}/posts")
-    public ResponseDto<?> getUserPosts(@PathVariable Long userSid, @PageableDefault(page = 0, size = 20) Pageable pageable) {
-        return ResponseDto.success(usersService.getUserPosts(userSid, pageable));
-    }
-    
     @GetMapping("/post/{sid}/user")
     public ResponseDto<?> findUser(@PathVariable Long sid) {
         return ResponseDto.success(usersService.findUserBySID(sid));
