+++ src/main/java/com/search/instagramsearching/service/UsersService.java
@@ -57,52 +57,6 @@ public class UsersService {
         return searchResultList;
     }
 
-    public List<UserPostsResponseDto> getUserPosts(Long userSid, Pageable pageable) {
-
-//        // profileName으로 profileId 찾기
-//        Long profileId = getProfileId(profileName, pageable);
-
-        // profileId로 해당 posts 조회하기
-        List<UserPostSearchResultDto> postsList = getSearchResult(userSid, pageable);
-        List<UserPostsResponseDto> response = new ArrayList<>();
-        for (UserPostSearchResultDto postInfo : postsList) {
-            response.add(
-                    UserPostsResponseDto.builder()
-                            .profileId(postInfo.getProfile_id())
-                            .profileName(null)
-                            .postId(postInfo.getPost_id())
-                            .locationId(postInfo.getLocation_id())
-                            .description(postInfo.getDescription())
-                            .cts(postInfo.getCts())
-                            .postType(postInfo.getPost_type())
-                            .numberLikes(postInfo.getNumber_likes())
-                            .numberComments(postInfo.getNumber_comments())
-                            .build()
-            );
-        }
-        return response;
-    }
-
-    // fulltext - profileNamed으로 users table 조회
-    private Long getProfileId(String profileName, Pageable pageable) {
-        List<UserSearchResultDto> searchResult = usersRepository.searchUsers(profileName, pageable);
-        if (searchResult.size() == 0) {
-            throw new UserNotFoundException();
-        }
-
-        Long profileId = searchResult.get(0).getProfile_id();
-        return profileId;
-    }
-
-    // Index - profileId로 posts table 조회
-    private List<UserPostSearchResultDto> getSearchResult(Long profileId, Pageable pageable) {
-        List<UserPostSearchResultDto> searchResult = postsRepository.getUserPosts(profileId, pageable);
-        if (searchResult.size() == 0) {
-            throw new PostsNotFoundExceptioin();
-        }
-        return searchResult;
-    }
-
     @Transactional
     public UserResponseDto findUserBySID (Long sid){
         Optional<Users> usersOptional = usersRepository.findUsersBySid(sid);
